refactor(utilities): refactor source of utility functions
