+++ src/index.js
@@ -5,12 +5,12 @@ import { resolve as resolvePath } from 'pathe';
 import { cwd as processCWD } from 'node:process';
 import { env as processEnv } from 'node:process';
 import { pathToFileURL } from 'node:url';
-import { isBuiltInModule } from './utilities.js';
-import { isInsideDir } from './utilities.js';
-import { isInsideNodeModules } from './utilities.js';
-import { resolveConfigFile } from './utilities.js';
-import { resolveProjectRoot } from './utilities.js';
-import { loadURL } from './utilities.js';
+import isBuiltInModule from './utilities/isBuiltInModule.js';
+import isInsideDir from './utilities/isInsideDir.js';
+import isInsideNodeModules from './utilities/isInsideNodeModules.js';
+import resolveConfigFile from './utilities/resolveConfigFile.js';
+import resolveProjectRoot from './utilities/resolveProjectRoot.js';
+import loadURL from './utilities/loadURL.js';
 
 const cwd = processCWD();
 const resolvedProjectRoot = await resolveProjectRoot( cwd );
