+++ src/utilities/isBuiltInModule.js
@@ -1,0 +1,9 @@
+function isBuiltInModule( { url, format } ) {
+	if ( format ) {
+		return format === 'builtin';
+	}
+
+	return url.startsWith( 'node:' );
+}
+
+export default isBuiltInModule;
