+++ src/utilities/isInsideNodeModules.js
@@ -1,0 +1,7 @@
+function isInsideNodeModules( pathOrURL ) {
+	const npmModulesPathRegex = /[/\\]node_modules[/\\]/gi;
+
+	return npmModulesPathRegex.test( pathOrURL );
+}
+
+export default isInsideNodeModules;
