+++ tests/__helpers__/fixtureDirPath.js
@@ -1,0 +1,8 @@
+import { dirname } from 'pathe';
+import { resolve as resolvePath } from 'pathe';
+import { fileURLToPath } from 'node:url';
+
+const __dirname = dirname( fileURLToPath( import.meta.url ) );
+const fixtureDirPath = resolvePath( __dirname, '..', '__fixtures__' );
+
+export default fixtureDirPath;
